# perf: improve equality comparison performance