--- a/lib/src/equatable_comparison_utils.dart
+++ b/lib/src/equatable_comparison_utils.dart
@@ -2,8 +2,34 @@ import 'package:equatable/src/equatable.dart';
 import 'package:equatable/src/equatable_mixin.dart';
 
 /// Determines whether [list1] and [list2] are equal.
+// This method is optimized for comparing properties
+// from primitive types like int, double, String, bool.
+// Using the _iterableEquals method instead of this code
+// can degrade performance by ~ 20%.
 bool equals(List<Object?> list1, List<Object?> list2) {
-  return _iterableEquals(list1, list2);
+  if (identical(list1, list2)) return true;
+  final length = list1.length;
+  if (length != list2.length) return false;
+
+  for (var i = 0; i < length; i++) {
+    final unit1 = list1[i];
+    final unit2 = list2[i];
+
+    if (_isEquatable(unit1) && _isEquatable(unit2)) {
+      if (unit1 != unit2) return false;
+    } else if (unit1 is Iterable && unit2 is Iterable) {
+      return _iterableEquals(unit1, unit2);
+    } else if (unit1 is Set && unit2 is Set) {
+      return _setEquals(unit1, unit2);
+    } else if (unit1 is Map && unit2 is Map) {
+      return _mapEquals(unit1, unit2);
+    } else if (unit1?.runtimeType != unit2?.runtimeType) {
+      return false;
+    } else if (unit1 != unit2) {
+      return false;
+    }
+  }
+  return true;
 }
 
 bool _objectsEquals(Object? object1, Object? object2) {