--- a/lib/src/equatable_mixin.dart
+++ b/lib/src/equatable_mixin.dart
@@ -1,5 +1,4 @@
 import 'package:equatable/src/equatable.dart';
-import 'package:equatable/src/equatable_comparison_utils.dart';
 import 'package:equatable/src/equatable_config.dart';
 import 'package:equatable/src/equatable_utils.dart';
 import 'package:meta/meta.dart';