## Status
**READY**

## Breaking Changes
NO

## Description
I've noticed that the performance of comparing objects with collections can be significantly enhanced. These optimizations in my application led to a considerable increase in FPS, approximately by 3 times. Additionally, I've [conducted tests](https://github.com/Maksimka101/equatable_benchmark) to compare the performance between the old and new approaches

## Todos
- [x] Tests
- [x] Documentation
- [x] Examples


## Impact to Remaining Code Base
This PR will affect: 
- comparison logic