@felangel 2023-12-11
Thanks for the PR! Are you able to share the benchmarks you used? I would love to run them locally and validate the results.