@escamoteur 2024-05-24
Just saw this. besides that I think it would definitely great to get this merged, why not borrow the optimization of `fast_equatable` and cache the hashcode if that really improves performance that much :-)