@felangel 2024-05-24
> Just saw this. besides that I think it would definitely great to get this merged, why not borrow the optimization of `fast_equatable` and cache the hashcode if that really improves performance that much :-)

Will review this later today apologies that it fell through the cracks. Looks like tests were failing but just re-triggered ci to see the logs.