@felangel 2024-05-26
> Just saw this. besides that I think it would definitely great to get this merged, why not borrow the optimization of `fast_equatable` and cache the hashcode if that really improves performance that much :-)

Because the classes should be immutable (e.g. all fields should be final and have a const constructor). Classes that use `package:fast_equatable` are not immutable and cannot have a const constructor.