@Maksimka101 2024-10-11
You are right. I found one bug in the `setEquals` function

Also, I've added the `@pragma('vm:prefer-inline')` annotatio to the `listEquals` function which improved performance by 20%-10% for some cases:

<details>

<summary>Benchmark details (when compiled to the exe file)</summary>

## With the inline pragma

```
EmptyEquatable
          total runs:  1 598 292   
          total time:     2.0000  s
         average run:          1 μs
         runs/second:  1 000 000   
               units:        100   
        units/second: 100 000 000   
       time per unit:     0.0100 μs

PrimitiveEquatable
          total runs:    678 313   
          total time:     2.0000  s
         average run:          2 μs
         runs/second:    500 000   
               units:        100   
        units/second: 50 000 000   
       time per unit:     0.0200 μs

CollectionEquatable (static, small)
          total runs:    139 323   
          total time:     2.0000  s
         average run:         14 μs
         runs/second:     71 429   
               units:        100   
        units/second:  7 142 857   
       time per unit:     0.1400 μs

CollectionEquatable (static, medium)
          total runs:    101 675   
          total time:     2.0000  s
         average run:         19 μs
         runs/second:     52 632   
               units:        100   
        units/second:  5 263 158   
       time per unit:     0.1900 μs

CollectionEquatable (static, large)
          total runs:     33 898   
          total time:     2.0000  s
         average run:         59 μs
         runs/second:     16 949   
               units:        100   
        units/second:  1 694 915   
       time per unit:     0.5900 μs

CollectionEquatable (dynamic, small)
          total runs:    458 845   
          total time:     2.0000  s
         average run:          4 μs
         runs/second:    250 000   
               units:        100   
        units/second: 25 000 000   
       time per unit:     0.0400 μs

CollectionEquatable (dynamic, medium)
          total runs:    460 127   
          total time:     2.0000  s
         average run:          4 μs
         runs/second:    250 000   
               units:        100   
        units/second: 25 000 000   
       time per unit:     0.0400 μs

CollectionEquatable (dynamic, large)
          total runs:    465 185   
          total time:     2.0000  s
         average run:          4 μs
         runs/second:    250 000   
               units:        100   
        units/second: 25 000 000   
       time per unit:     0.0400 μs
```

## Without the inline pragma

```
EmptyEquatable
          total runs:  1 332 271   
          total time:     2.0000  s
         average run:          1 μs
         runs/second:  1 000 000   
               units:        100   
        units/second: 100 000 000   
       time per unit:     0.0100 μs

PrimitiveEquatable
          total runs:    585 970   
          total time:     2.0000  s
         average run:          3 μs
         runs/second:    333 333   
               units:        100   
        units/second: 33 333 333   
       time per unit:     0.0300 μs

CollectionEquatable (static, small)
          total runs:    133 655   
          total time:     2.0000  s
         average run:         14 μs
         runs/second:     71 429   
               units:        100   
        units/second:  7 142 857   
       time per unit:     0.1400 μs

CollectionEquatable (static, medium)
          total runs:     99 270   
          total time:     2.0000  s
         average run:         20 μs
         runs/second:     50 000   
               units:        100   
        units/second:  5 000 000   
       time per unit:     0.2000 μs

CollectionEquatable (static, large)
          total runs:     33 109   
          total time:     2.0001  s
         average run:         60 μs
         runs/second:     16 667   
               units:        100   
        units/second:  1 666 667   
       time per unit:     0.6000 μs

CollectionEquatable (dynamic, small)
          total runs:    423 572   
          total time:     2.0000  s
         average run:          4 μs
         runs/second:    250 000   
               units:        100   
        units/second: 25 000 000   
       time per unit:     0.0400 μs

CollectionEquatable (dynamic, medium)
          total runs:    410 959   
          total time:     2.0000  s
         average run:          4 μs
         runs/second:    250 000   
               units:        100   
        units/second: 25 000 000   
       time per unit:     0.0400 μs

CollectionEquatable (dynamic, large)
          total runs:    417 545   
          total time:     2.0000  s
         average run:          4 μs
         runs/second:    250 000   
               units:        100   
        units/second: 25 000 000   
       time per unit:     0.0400 μs
```

</details>